no need for "double" in GetAverageAnonymizedRounds, "float" should be enough

+++ src/qt/overviewpage.cpp
@@ -351,7 +351,7 @@ void OverviewPage::updatePrivateSendProgress()
     CAmount nDenominatedUnconfirmedBalance;
     CAmount nAnonymizableBalance;
     CAmount nNormalizedAnonymizedBalance;
-    double nAverageAnonymizedRounds;
+    float nAverageAnonymizedRounds;
 
     {
         nDenominatedConfirmedBalance = pwalletMain->GetDenominatedBalance();
